Differentiate between a directory that exists, and one that's also valid
FileSystemItem keeps track of whether it's pointed at a valid directory (no errors upon opening and reading), and is_directory returned the current validity. is_directory has been updated to return if a path is a directory, regardless of whether it's valid--it only cares if it exists. Instead of using is_directory internally as it did before, FileSystemItem now validates a directory using its already existing private variable, which is updated anytime the path is changed.