Add the HSplitRatioContainer node
Inherits from a regular HSplitContainer; this node allows the user to set a specific size ratio between the left and right sides of the container. The ratio will be dynamically updated when the splitter is dragged, and the current ratio will be maintained if the container is resized (such as when resizing the window).